+++ example/lib/main.dart
@@ -11,7 +11,10 @@ import 'package:example/component_list/pp_inbox.dart';
 import 'package:example/component_list/grid_component.dart';
 
 import 'package:example/component_list/typography_component.dart';
+import 'package:example/revision/backdrop/backdrop_screen.dart';
+import 'package:example/revision/font/font_screen.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_direct/flutter_direct.dart';
 
 void main() => runApp(MyApp());
 
@@ -21,17 +24,19 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       initialRoute: '/',
       routes: {
-        '/button': (context) => ButtonComponent(),
-        '/typography': (context) => TypographyComponent(),
-        '/color': (context) => ColorComponent(),
-        '/input': (context) => InputComponent(),
-        '/listtile': (context) => ListitleComponent(),
-        '/homepp': (context) => PPHomeScreen(),
-        '/homeppv5': (context) => PPHomeScreenV5(),
-        '/homeppv6': (context) => PPHomeScreenV6(),
-        '/inboxpp': (context) => PPInboxScreen(),
-        '/image': (context) => ImageComponent(),
-        '/grid': (context) => GridComponent()
+        // '/button': (context) => ButtonComponent(),
+        // '/typography': (context) => TypographyComponent(),
+        // '/color': (context) => ColorComponent(),
+        // '/input': (context) => InputComponent(),
+        // '/listtile': (context) => ListitleComponent(),
+        // '/homepp': (context) => PPHomeScreen(),
+        // '/homeppv5': (context) => PPHomeScreenV5(),
+        // '/homeppv6': (context) => PPHomeScreenV6(),
+        // '/inboxpp': (context) => PPInboxScreen(),
+        // '/image': (context) => ImageComponent(),
+        // '/grid': (context) => GridComponent()
+        '/font': (context) => FontScreen(),
+        '/backdrop': (context) => BackdropScreen()
       },
       theme: ThemeData(primaryColor: Colors.white),
       home: HomeScreen(),
@@ -59,81 +64,88 @@ class _HomeScreenState extends State<HomeScreen> {
                 children: <Widget>[
                   ListTile(
                     trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Typography Component"),
+                    title: Text("Font Component"),
                     onTap: () {
-                      Navigator.pushNamed(context, '/typography');
+                      Navigator.pushNamed(context, '/font');
                     },
                   ),
                   ListTile(
                     trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Color Component"),
+                    title: Text("Backdrop Component"),
                     onTap: () {
-                      Navigator.pushNamed(context, '/color');
+                      Navigator.pushNamed(context, '/backdrop');
                     },
                   ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Button Component"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/button');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Input Component"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/input');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("List Tile Component"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/listtile');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Home Partner Portal"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/homepp');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Home Partner Portal V5"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/homeppv5');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Home Partner Portal V6"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/homeppv6');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Inbox Partner Portal"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/inboxpp');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Image Component"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/image');
-                    },
-                  ),
-                  ListTile(
-                    trailing: Icon(Icons.arrow_forward_ios),
-                    title: Text("Grid Component"),
-                    onTap: () {
-                      Navigator.pushNamed(context, '/grid');
-                    },
-                  )
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Color Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/color');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Button Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/button');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Input Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/input');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("List Tile Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/listtile');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Home Partner Portal"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/homepp');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Home Partner Portal V5"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/homeppv5');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Home Partner Portal V6"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/homeppv6');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Inbox Partner Portal"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/inboxpp');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Image Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/image');
+                  //   },
+                  // ),
+                  // ListTile(
+                  //   trailing: Icon(Icons.arrow_forward_ios),
+                  //   title: Text("Grid Component"),
+                  //   onTap: () {
+                  //     Navigator.pushNamed(context, '/grid');
+                  //   },
+                  // )
                 ],
               ),
             )));