+++ lib/revision/commons/atoms/font/atomic_font.dart
@@ -1,0 +1,7 @@
+part of flutter_direct;
+
+class AtomicFont {
+  static const String Helvetica = "HelveticaNeue";
+  static const String Raleway = "Raleway";
+  static const String Campton = "Campton";
+}